cmdmon: use 32-bit fields in client access report
The clientlog record still uses 16-bit integers to count dropped
packets, but this will avoid an incompatible change in the command
reply if there will be a need to count more than 2^16 drops.

+++ candm.h
@@ -553,8 +553,8 @@ typedef struct {
   IPAddr ip;
   uint32_t ntp_hits;
   uint32_t cmd_hits;
-  uint16_t ntp_drops;
-  uint16_t cmd_drops;
+  uint32_t ntp_drops;
+  uint32_t cmd_drops;
   int8_t ntp_interval;
   int8_t cmd_interval;
   int8_t ntp_timeout_interval;
